Implement extract-on-open.
Add a CContextMenuInfo setting to this effect.

Add an FM switch "-open" in the default file type handler command. Call
ExtractArchives if this switch is specified and the setting is enabled.

When using extract-on-open, always show the result folder. This is
always necessary because opening archives from some apps will not reveal
the extracted folder otherwise.

Signed-off-by: Tu Dinh <contact@tudinh.xyz>