Implement extract-on-open. (#709)
* Pass the real OutDir to OpenFolder.

In cases where the OutDir is not specified (e.g. smart extract),
ShellExecuteW will be called with the wrong directory path.

To resolve this issue, pass the real OutDir through this route:
* Add CArchiveExtractCallback.OutDir, which is set by DecompressArchive;
* At the end of Extract(), copy CArchiveExtractCallback.OutDir to
  CDecompressStat.OutDir;
* Make CDecompressStat a member of CExtractCallbackImp;
* Obtain CDecompressStat.OutDir via CExtractCallbackImp.Stat at the end.

Signed-off-by: Tu Dinh <contact@tudinh.xyz>

* Add -sre option to open folder after extracting.

Signed-off-by: Tu Dinh <contact@tudinh.xyz>

* Implement extract-on-open.

Add a CContextMenuInfo setting to this effect.

Add an FM switch "-open" in the default file type handler command. Call
ExtractArchives if this switch is specified and the setting is enabled.

When using extract-on-open, always show the result folder. This is
always necessary because opening archives from some apps will not reveal
the extracted folder otherwise.

Signed-off-by: Tu Dinh <contact@tudinh.xyz>

---------

Signed-off-by: Tu Dinh <contact@tudinh.xyz>